requirements.txt
gspread
google-auth
pandas
openpyxl
pydrive
oauth2client
google-cloud-vision
Pillow